improve likely_html check
